Implement ShortUrl model
I chose to tackle the model first, because the rest of the functionality
will stem naturally from there. My approach was to sequentially make
each test in short_url_spec.rb pass.

From the test spec, I was able to see that the short_code string is a
base-62 representation of the ShortUrl ID, 62 being the number of
alphanumeric characters. That makes the short_code and find_by_short_code
methods simply encoding/decoding the ID.